refactor extracted code such that it also works with managed system
