+++ bundles/org.palladiosimulator.analyzer.slingshot.managedsystem.data/META-INF/MANIFEST.MF
@@ -11,9 +11,8 @@ Require-Bundle: org.palladiosimulator.spd,
  org.palladiosimulator.servicelevelobjective,
  org.palladiosimulator.edp2.util,
  org.palladiosimulator.analyzer.slingshot.core,
- org.apache.log4j
-Export-Package: org.palladiosimulator.analyzer.slingshot.managedsystem.converter,
- org.palladiosimulator.analyzer.slingshot.managedsystem.data,
- org.palladiosimulator.analyzer.slingshot.managedsystem.events,
+ org.apache.log4j,
+ org.palladiosimulator.analyzer.slingshot.stateexploration.converter
+Export-Package: org.palladiosimulator.analyzer.slingshot.managedsystem.events,
  org.palladiosimulator.analyzer.slingshot.managedsystem.messages,
  org.palladiosimulator.analyzer.slingshot.managedsystem.messages.data
