--- bundles/org.palladiosimulator.analyzer.slingshot.managedsystem.data/src/org/palladiosimulator/analyzer/slingshot/managedsystem/data/SLO.java
@@ -1,5 +1,0 @@
-package org.palladiosimulator.analyzer.slingshot.managedsystem.data;
-
-public record SLO(String id, String name, String specificationId, Number lowerThreshold, Number upperThreshold) {
-
-}
