+++ bundles/org.palladiosimulator.analyzer.slingshot.managedsystem/src/org/palladiosimulator/analyzer/slingshot/managedsystem/SendMessagesBehaviour.java
@@ -10,6 +10,8 @@ import javax.inject.Named;
 import org.apache.log4j.Logger;
 import org.palladiosimulator.analyzer.slingshot.common.annotations.Nullable;
 import org.palladiosimulator.analyzer.slingshot.common.events.DESEvent;
+import org.palladiosimulator.analyzer.slingshot.converter.StateGraphConverter;
+import org.palladiosimulator.analyzer.slingshot.converter.data.StateGraphNode;
 import org.palladiosimulator.analyzer.slingshot.core.Slingshot;
 import org.palladiosimulator.analyzer.slingshot.core.api.SystemDriver;
 import org.palladiosimulator.analyzer.slingshot.core.events.PreSimulationConfigurationStarted;
@@ -19,8 +21,6 @@ import org.palladiosimulator.analyzer.slingshot.core.extension.SimulationBehavio
 import org.palladiosimulator.analyzer.slingshot.eventdriver.annotations.Subscribe;
 import org.palladiosimulator.analyzer.slingshot.eventdriver.annotations.eventcontract.OnEvent;
 import org.palladiosimulator.analyzer.slingshot.eventdriver.returntypes.Result;
-import org.palladiosimulator.analyzer.slingshot.managedsystem.converter.StateGraphConverter;
-import org.palladiosimulator.analyzer.slingshot.managedsystem.data.StateGraphNode;
 import org.palladiosimulator.analyzer.slingshot.managedsystem.events.UtilityIntervalPassed;
 import org.palladiosimulator.analyzer.slingshot.managedsystem.messages.ManagedSystemFinishedMessage;
 import org.palladiosimulator.analyzer.slingshot.managedsystem.messages.ManagedSystemStartedMessage;
@@ -142,8 +142,8 @@ public class SendMessagesBehaviour implements SimulationBehaviorExtension {
     private void publishUtility(final DESEvent event) {
         assert this.expSetting != null : "ExperimentSettings are not yet set.";
 
-        final StateGraphNode node = StateGraphConverter.convertState(this.monitorRepo, this.expSetting, this.sloRepo,
-                prevUtilityTimestamp, event.time());
+        final StateGraphNode node = StateGraphConverter.convertState(Optional.of(this.monitorRepo), this.expSetting, Optional.of(this.sloRepo),
+                prevUtilityTimestamp, event.time(), "", "", List.of());
 
         this.systemDriver.postEvent(new StateExploredEventMessage(node, clientName));
     }
