--- bundles/org.palladiosimulator.analyzer.slingshot.stateexploration.converter.data/META-INF/MANIFEST.MF
@@ -1,12 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: StateexplorationConverterData
-Bundle-SymbolicName: org.palladiosimulator.analyzer.slingshot.stateexploration.converter.data
-Bundle-Version: 1.0.0.qualifier
-Automatic-Module-Name: org.palladiosimulator.analyzer.slingshot.stateexploration.converter.data
-Bundle-RequiredExecutionEnvironment: JavaSE-17
-Export-Package: org.palladiosimulator.analyzer.slingshot.converter.data
-Require-Bundle: org.jscience,
- org.palladiosimulator.servicelevelobjective,
- org.palladiosimulator.spd,
- org.apache.log4j
