--- bundles/org.palladiosimulator.analyzer.slingshot.stateexploration.converter.data/build.properties
@@ -1,4 +1,0 @@
-source.. = src/
-output.. = bin/
-bin.includes = META-INF/,\
-               .
