+++ bundles/org.palladiosimulator.analyzer.slingshot.stateexploration.converter/src/org/palladiosimulator/analyzer/slingshot/converter/data/SLO.java
